requirements.txt
absl-py==0.8.1
Adafruit-CharLCD==1.1.1
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.3
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.13
decorator==4.0.11
defusedxml==0.6.0
Django==2.0.1
entrypoints==0.3
filetype==1.0.5
future==0.18.1
gast==0.3.2
google-pasta==0.1.7
googletrans==2.4.0
grpcio==1.24.3
gTTS==2.0.4
gTTS-token==1.1.3
h5py==2.10.0
idna==2.6
imageai==2.1.5
imageio==2.1.2
importlib-metadata==0.23
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
InstagramAPI==1.0.2
instapy-cli==0.0.12
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
moviepy==0.2.3.2
nbconvert==5.6.0
nbformat==4.4.0
netifaces==0.10.9
notebook==6.0.1
numpy==1.16.2
oauthlib==3.1.0
opencv-python==3.4.3.18
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
playsound==1.2.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pycrypto==2.6.1
pygame==1.9.6
Pygments==2.4.2
PyGObject==3.30.4
pyparsing==2.4.2
pyrsistent==0.15.4
pyserial==3.4
PySocks==1.7.1
python-apt==1.8.4
python-dateutil==2.8.0
python-telegram-bot==12.2.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.11.1
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
RPi.GPIO==0.7.0
scipy==1.3.1
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.4
spidev==3.4
sqlparse==0.3.0
ssh-import-id==5.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.11.2
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
